Remove src/runtime-test-bench/: extract live files, delete orphaned subsystem (#521)
* Initial plan

* feat: move runtime-test-bench live files to proper locations and delete orphaned code

Agent-Logs-Url: https://github.com/SergeiGolos/wod-wiki/sessions/4dcb4cb2-71ad-4b70-b0ab-ba586cb342e2


---------
