+++ src/components/Editor/overlays/RuntimeTimerPanel.tsx
@@ -25,8 +25,8 @@ import { TimerDisplay } from "@/panels/timer-panel";
 import { VisualStatePanel } from "@/panels/visual-state-panel";
 import { PanelSizeProvider, usePanelSize } from "@/panels/panel-system/PanelSizeContext";
 import { RuntimeFactory } from "@/runtime/compiler/RuntimeFactory";
-import { globalCompiler, globalParser } from "@/runtime-test-bench/services/testbench-services";
-import { useRuntimeExecution, type UseRuntimeExecutionReturn } from "@/runtime-test-bench/hooks/useRuntimeExecution";
+import { globalCompiler, globalParser } from "@/runtime/services/runtimeServices";
+import { useRuntimeExecution, type UseRuntimeExecutionReturn } from "@/runtime/hooks/useRuntimeExecution";
 import { useWorkbenchSyncStore } from "@/components/layout/workbenchSyncStore";
 import { SubscriptionManager } from "@/runtime/subscriptions/SubscriptionManager";
 import { ChromecastRuntimeSubscription } from "@/services/cast/rpc/ChromecastRuntimeSubscription";
