+++ src/components/debug/MemoryValuePopover.tsx
