+++ src/index.ts
@@ -33,6 +33,3 @@ export * from './clock-entry';
 
 // Styles
 import './index.css';
-
-
-// export * from './runtime-test-bench';
