--- src/runtime-test-bench/adapters/index.ts
@@ -1,1 +1,0 @@
-export { RuntimeAdapter } from './RuntimeAdapter';
