--- src/runtime-test-bench/components/index.ts
@@ -1,7 +1,0 @@
-export { EditorPanel } from './EditorPanel';
-export { RuntimeStackPanel } from './RuntimeStackPanel';
-export { MemoryPanel } from './MemoryPanel';
-export { Toolbar } from './Toolbar';
-export { CompilationPanel } from './CompilationPanel';
-export { ControlsPanel } from './ControlsPanel';
-export { StatusFooter } from './StatusFooter';
