--- src/runtime-test-bench/index.ts
@@ -1,3 +1,0 @@
-export * from './components';
-export * from './types/interfaces';
-export * from './styles/tailwind-components';
