+++ stories/catalog/templates/Tracker/Web.stories.tsx
@@ -49,7 +49,7 @@ import { PanelSizeProvider } from '@/panels/panel-system/PanelSizeContext';
 import { DebugModeProvider } from '@/components/layout/DebugModeContext';
 import { VisualStatePanel } from '@/panels/visual-state-panel';
 import { TimerDisplay } from '@/panels/timer-panel';
-import { useRuntimeExecution } from '@/runtime-test-bench/hooks/useRuntimeExecution';
+import { useRuntimeExecution } from '@/runtime/hooks/useRuntimeExecution';
 // ─────────────────────────────────────────────────────────────────────────────
 // Helpers
 // ─────────────────────────────────────────────────────────────────────────────
